Add IdManager

+++ CasaEngine/Framework/AI/Messaging/IMessageManager.cs
@@ -5,7 +5,7 @@ public interface IMessageManager
 
     void ResetManager(double precision);
 
-    void SendMessage(int senderId, int recieverId, double delayTime, int type, object extraInfo);
+    void SendMessage(long senderId, long recieverId, double delayTime, int type, object extraInfo);
 
     void Update();
